+++ frontend/src/models/AuthToken.component.ts
@@ -1,0 +1,8 @@
+export interface AuthToken {
+    exp: number
+    firstName: string
+    lastName: string
+    username: string
+    iat: number
+    iss: string
+}
