+++ frontend/src/models/LoginRequest.model.ts
@@ -1,0 +1,3 @@
+export class LoginRequest {
+    constructor(readonly username: string, readonly password: string){}
+}
